+++ lib/main.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
+import 'core/constants/app_constants.dart';
 import 'core/theme/app_theme.dart';
 import 'routes/app_pages.dart';
 import 'routes/app_routes.dart';
@@ -15,7 +16,7 @@ class KanzooApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return GetMaterialApp(
-      title: 'Kanzoo',
+      title: AppConstants.appName,
       theme: AppTheme.lightTheme,
       initialRoute: AppRoutes.splash,
       getPages: AppPages.routes,